feat(stepper): allow fractional gear_ratio + enforce max_step_vel cap
Two related changes to make the stepper driver robust against
mis-configuration and non-integer gearboxes:

1. gear_ratio_ becomes double everywhere (wrapper member, StepperDriver
   ctor + member). Previously declared as int, which silently truncated
   any non-integer URDF value (e.g. a 5.18:1 worm gear would have been
   treated as 5:1, ~3.5% velocity error).

2. StepperDriver::set_velocity now clamps the requested microsteps/s
   at max_step_vel_ (and at uint16_t range) instead of letting it
   overflow or wildly overshoot the configured cap. Without the clamp
   the max_step_vel parameter in the launch config was effectively
   dead — silent_stepper_v2_set_max_velocity was set once at init and
   then unconditionally overwritten on every cmd_vel callback.

+++ include/cmeresearch_stepper_driver/stepper_driver_ros_wrapper.hpp
@@ -79,7 +79,7 @@ private:
     int deceleration_;
     int steps_per_revolution_;
     bool mirror_direction_;
-    int gear_ratio_;
+    double gear_ratio_;
     int max_step_vel_;
     std::string wheel_name_;
     bool hw_simulation_;
